Extract handlePaymentLink logic from Dashboard to new blank screen HandlePaymentLink
